move albu train augs to pipelines
